requirements.txt
keras
wandb
tensorflow
pillow
opencv-python
scikit-image
imgaug
